refactor: extract danger button logic
