+++ projects/client/src/lib/components/buttons/_internal/useDangerButton.ts
@@ -1,20 +1,24 @@
 import { useMedia, WellKnownMediaQuery } from '$lib/stores/css/useMedia';
 import { onMount } from 'svelte';
 import { derived, get, writable } from 'svelte/store';
+import type { TraktButtonProps } from '../TraktButtonProps';
 
+type ButtonColor = Exclude<TraktButtonProps['color'], Nil>;
 type DangerProps = {
+  color: ButtonColor;
   isActive: boolean;
 };
-export function useDangerColor({
+export function useDangerButton({
+  color: seed,
   isActive,
 }: DangerProps) {
-  const stateToColor = (isActive: boolean) => (isActive ? 'red' : 'purple');
+  const stateToColor = (isActive: boolean) => (isActive ? 'red' : seed);
   const interactionToColor = (isTouch: boolean, isActive: boolean) =>
-    isTouch ? stateToColor(isActive) : 'purple';
+    isTouch ? stateToColor(isActive) : seed;
 
   const isTouch = useMedia(WellKnownMediaQuery.touch);
 
-  const color = writable<'purple' | 'red'>(
+  const color = writable<ButtonColor>(
     interactionToColor(get(isTouch), isActive),
   );
 
@@ -27,6 +31,11 @@ export function useDangerColor({
 
   return {
     color,
+    variant: derived(
+      isTouch,
+      ($isTouch) => isActive && !$isTouch ? 'primary' : 'secondary',
+    ),
+    isTouch,
     onmouseover: () => color.set(stateToColor(isActive)),
     onfocusin: () => color.set(stateToColor(isActive)),
     onfocusout: () => color.set(interactionToColor(get(isTouch), isActive)),
